# openapi.yaml
openapi: 3.1.0
info:
  title: Herald API
  description: Persistent realtime event streams with built-in authorization.
  license:
    name: Proprietary
  version: 2.0.0
servers:
- url: /
  description: Local server
paths:
  /admin/connections:
    get:
      tags:
      - admin
      summary: List active connections.
      operationId: list_connections
      responses:
        '200':
          description: Active connections
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionListResponse'
      security:
      - admin_auth: []
  /admin/errors:
    get:
      tags:
      - admin
      summary: List recent errors.
      operationId: list_errors
      parameters:
      - name: category
        in: query
        required: false
        schema:
          type:
          - string
          - 'null'
      - name: limit
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          minimum: 0
      responses:
        '200':
          description: Recent errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorListResponse'
      security:
      - admin_auth: []
  /admin/events:
    get:
      tags:
      - admin
      summary: List recent admin events.
      operationId: list_events
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          minimum: 0
      - name: after_id
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int64
          minimum: 0
      responses:
        '200':
          description: Recent admin events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminEventListResponse'
      security:
      - admin_auth: []
  /admin/events/stream:
    get:
      tags:
      - admin
      summary: SSE stream of admin events.
      operationId: events_stream
      responses:
        '200':
          description: SSE event stream
          content:
            text/event-stream: {}
      security:
      - admin_auth: []
  /admin/stats:
    get:
      tags:
      - admin
      summary: Get stats snapshots and today's summary.
      operationId: get_stats
      parameters:
      - name: from
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int64
      - name: to
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int64
      responses:
        '200':
          description: Stats snapshots
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminStatsResponse'
      security:
      - admin_auth: []
  /admin/tenants:
    get:
      tags:
      - admin
      operationId: list_tenants
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          minimum: 0
      - name: offset
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          minimum: 0
      responses:
        '200':
          description: List of tenants
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantListResponse'
      security:
      - admin_auth: []
    post:
      tags:
      - admin
      operationId: create_tenant
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenantRequest'
        required: true
      responses:
        '201':
          description: Tenant created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantCreateResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Tenant creation conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - admin_auth: []
  /admin/tenants/{id}:
    get:
      tags:
      - admin
      operationId: get_tenant
      parameters:
      - name: id
        in: path
        description: Tenant ID
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Tenant details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantGetResponse'
        '404':
          description: Tenant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - admin_auth: []
    delete:
      tags:
      - admin
      operationId: delete_tenant
      parameters:
      - name: id
        in: path
        description: Tenant ID
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Tenant deleted
        '404':
          description: Tenant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - admin_auth: []
    patch:
      tags:
      - admin
      operationId: update_tenant
      parameters:
      - name: id
        in: path
        description: Tenant ID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTenantRequest'
        required: true
      responses:
        '200':
          description: Tenant updated
        '404':
          description: Tenant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - admin_auth: []
  /admin/tenants/{id}/audit:
    get:
      tags:
      - admin
      operationId: query_audit
      parameters:
      - name: id
        in: path
        description: Tenant ID
        required: true
        schema:
          type: string
      - name: resource_type
        in: query
        required: false
        schema:
          type:
          - string
          - 'null'
      - name: resource_id
        in: query
        required: false
        schema:
          type:
          - string
          - 'null'
      - name: actor
        in: query
        required: false
        schema:
          type:
          - string
          - 'null'
      - name: operation
        in: query
        required: false
        schema:
          type:
          - string
          - 'null'
      - name: result
        in: query
        required: false
        schema:
          type:
          - string
          - 'null'
      - name: since
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int64
          minimum: 0
      - name: until
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int64
          minimum: 0
      - name: limit
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          minimum: 0
      responses:
        '200':
          description: Audit log entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditQueryResponse'
        '501':
          description: Audit not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - admin_auth: []
  /admin/tenants/{id}/audit/count:
    get:
      tags:
      - admin
      operationId: count_audit
      parameters:
      - name: id
        in: path
        description: Tenant ID
        required: true
        schema:
          type: string
      - name: resource_type
        in: query
        required: false
        schema:
          type:
          - string
          - 'null'
      - name: resource_id
        in: query
        required: false
        schema:
          type:
          - string
          - 'null'
      - name: actor
        in: query
        required: false
        schema:
          type:
          - string
          - 'null'
      - name: operation
        in: query
        required: false
        schema:
          type:
          - string
          - 'null'
      - name: result
        in: query
        required: false
        schema:
          type:
          - string
          - 'null'
      - name: since
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int64
          minimum: 0
      - name: until
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int64
          minimum: 0
      - name: limit
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          minimum: 0
      responses:
        '200':
          description: Audit event count
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditCountResponse'
        '501':
          description: Audit not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - admin_auth: []
  /admin/tenants/{id}/data:
    delete:
      tags:
      - admin
      summary: 'GDPR: Purge ALL data for a tenant from all namespaces.'
      operationId: purge_tenant_data
      parameters:
      - name: id
        in: path
        description: Tenant ID
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Tenant data purged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurgeResponse'
        '404':
          description: Tenant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - admin_auth: []
  /admin/tenants/{id}/streams:
    get:
      tags:
      - admin
      summary: List streams for a specific tenant (admin view).
      operationId: list_tenant_streams
      parameters:
      - name: id
        in: path
        description: Tenant ID
        required: true
        schema:
          type: string
      - name: limit
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          minimum: 0
      - name: offset
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          minimum: 0
      responses:
        '200':
          description: Tenant streams
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreamListResponse'
      security:
      - admin_auth: []
  /admin/tenants/{id}/tokens:
    get:
      tags:
      - admin
      operationId: list_api_tokens
      parameters:
      - name: id
        in: path
        description: Tenant ID
        required: true
        schema:
          type: string
      - name: limit
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          minimum: 0
      - name: offset
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          minimum: 0
      responses:
        '200':
          description: List of API tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenListResponse'
      security:
      - admin_auth: []
    post:
      tags:
      - admin
      operationId: create_api_token
      parameters:
      - name: id
        in: path
        description: Tenant ID
        required: true
        schema:
          type: string
      requestBody:
        description: Optional token scope
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTokenRequest'
        required: true
      responses:
        '201':
          description: API token created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenCreateResponse'
        '404':
          description: Tenant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - admin_auth: []
  /admin/tenants/{id}/tokens/{token}:
    delete:
      tags:
      - admin
      operationId: delete_api_token
      parameters:
      - name: id
        in: path
        description: Tenant ID
        required: true
        schema:
          type: string
      - name: token
        in: path
        description: API token
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Token deleted
        '404':
          description: Token not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - admin_auth: []
...
  /presence/{user_id}:
    get:
      tags:
      - presence
      operationId: user_presence
      parameters:
      - name: user_id
        in: path
        description: User ID
        required: true
        schema:
          type: string
      responses:
        '200':
          description: User presence status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPresenceResponse'
      security:
      - basic_auth: []
      - bearer_auth: []
    post:
      tags:
      - presence
      operationId: admin_set_presence
      parameters:
      - name: user_id
        in: path
        description: User ID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminSetPresenceRequest'
        required: true
      responses:
        '200':
          description: Presence override applied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPresenceResponse'
      security:
      - basic_auth: []
      - bearer_auth: []
  /self/connections:
    get:
      tags:
      - self-service
      summary: GET /self/connections — tenant's active connection count.
      operationId: list_connections
      responses:
        '200':
          description: Tenant connection count
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SelfConnectionResponse'
      security:
      - basic_auth: []
      - bearer_auth: []
  /self/errors:
    get:
      tags:
      - self-service
      summary: GET /self/errors — tenant's recent errors.
      operationId: list_errors
      parameters:
      - name: category
        in: query
        required: false
        schema:
          type:
          - string
          - 'null'
      - name: limit
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          minimum: 0
      responses:
        '200':
          description: Tenant errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorListResponse'
      security:
      - basic_auth: []
      - bearer_auth: []
  /self/events:
    get:
      tags:
      - self-service
      summary: GET /self/events — tenant's recent admin events.
      operationId: list_events
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          minimum: 0
      - name: after_id
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int64
          minimum: 0
      responses:
        '200':
          description: Tenant admin events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminEventListResponse'
      security:
      - basic_auth: []
      - bearer_auth: []
  /self/events/stream:
    get:
      tags:
      - self-service
      summary: GET /self/events/stream — SSE stream of tenant's events.
      operationId: events_stream
      responses:
        '200':
          description: SSE event stream
          content:
            text/event-stream: {}
      security:
      - basic_auth: []
      - bearer_auth: []
  /self/secret/rotate:
    post:
      tags:
      - self-service
      summary: POST /self/secret/rotate — rotate the tenant's secret, return new credentials.
      operationId: rotate_secret
      responses:
        '200':
          description: Secret rotated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RotateSecretResponse'
        '404':
          description: Tenant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - basic_auth: []
      - bearer_auth: []
  /self/tokens:
    get:
      tags:
      - self-service
      summary: GET /self/tokens — list API tokens for the authenticated tenant.
      operationId: list_tokens
      responses:
        '200':
          description: List of tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SelfTokenListResponse'
      security:
      - basic_auth: []
      - bearer_auth: []
    post:
      tags:
      - self-service
      summary: POST /self/tokens — create an API token for the authenticated tenant.
      operationId: create_token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTokenRequest'
        required: true
      responses:
        '201':
          description: Token created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenCreateResponse'
      security:
      - basic_auth: []
      - bearer_auth: []
...
components:
  schemas:
...
    AdminEventListResponse:
      type: object
      required:
      - events
      properties:
        events:
          type: array
          items: {}
...
    AdminStatsResponse:
      type: object
      required:
      - today
      - snapshots
      properties:
        snapshots:
          type: array
          items: {}
        today:
          $ref: '#/components/schemas/TodaySummary'
...
    ConnectionListResponse:
      type: object
      required:
      - total
      - by_tenant
      properties:
        by_tenant:
          type: array
          items:
            $ref: '#/components/schemas/TenantConnectionCount'
        total:
          type: integer
          minimum: 0
...
    CreateTenantRequest:
      type: object
      required:
      - name
      properties:
        name:
          type: string
        plan:
          type:
          - string
          - 'null'
...
    ErrorListResponse:
      type: object
      required:
      - errors
      properties:
        errors:
          type: array
          items: {}
...
    TenantListResponse:
      type: object
      required:
      - tenants
      - total
      - limit
      - offset
      properties:
        limit:
          type: integer
          minimum: 0
        offset:
          type: integer
          minimum: 0
        tenants:
          type: array
          items:
            $ref: '#/components/schemas/TenantListItem'
        total:
          type: integer
          minimum: 0
...
  securitySchemes:
    admin_auth:
      type: http
      scheme: bearer
    basic_auth:
      type: http
      scheme: basic
    bearer_auth:
      type: http
      scheme: bearer
security:
- basic_auth: []
- bearer_auth: []
- admin_auth: []
...